+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to Replicator will be documented in this file.
 >The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Refactored VariantPool into an abstract class that can be used as a base for other multi-pool strategies
+
 ## [0.3.0-alpha]
 
 ### Added